% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-UserBayesFunctions.R
\name{bayes.update}
\alias{bayes.update}
\title{Updating an Object of Class \code{minimax}}
\usage{
bayes.update(object, iter, ...)
}
\arguments{
\item{object}{An object of class \code{minimax}.}

\item{iter}{Number of iterations.}

\item{...}{An argument of no further use.}
}
\description{
Runs the ICA optimization algorithm on an object of class \code{minimax} for more number of iterations  and updates the results.
}
\seealso{
\code{\link{bayes}}
}
