% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FIM_sig_emax}
\alias{FIM_sig_emax}
\title{Fisher information matrix for the sigmoid emax model.}
\usage{
FIM_sig_emax(x, w, param)
}
\arguments{
\item{x}{vector of design points.}

\item{w}{vector of design weight. Its length must be equal to the length of \code{x} and \code{sum(w)} should be 1.}

\item{param}{vector of model parameters \eqn{(\theta_1, \theta_2,\theta_3, \theta_4)}{(\theta1, \theta2,\theta3, \theta4)}.}
}
\value{
Fisher information matrix of design.
}
\description{
The mean of response variable is .
}
