% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-UserBayesFunctions.R
\name{iterate.bayes}
\alias{iterate.bayes}
\title{Updating an Object of Class \code{bayes}}
\usage{
\method{iterate}{bayes}(object, iter)
}
\arguments{
\item{object}{An object of class \code{bayes}.}

\item{iter}{Number of iterations.}
}
\description{
Runs the ICA optimization algorithm on an object of class \code{bayes} for more number of iterations  and updates the results.
}
\seealso{
\code{\link{bayes}}
}
