% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_iterate.R
\name{iterate.ICA}
\alias{iterate.ICA}
\title{Update an object of class 'ICA'}
\usage{
\method{iterate}{ICA}(object, iter)
}
\arguments{
\item{object}{an object of class 'ICA'.}

\item{iter}{number of iterations.}
}
\description{
Update an object of class 'ICA'
}
\seealso{
\code{\link{mica}}
}

