% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-UserMinimaxFunctions.R
\name{iterate.minimax}
\alias{iterate.minimax}
\title{Updating an Object of Class \code{minimax}}
\usage{
\method{iterate}{minimax}(object, iter)
}
\arguments{
\item{object}{An object of class \code{minimax}.}

\item{iter}{Number of iterations.}
}
\description{
Runs the ICA optimization algorithm on an object of class \code{minimax} for more number of iterations  and updates the results.
}
\seealso{
\code{\link{minimax}}
}
