% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_plot.R
\name{plot.ICA}
\alias{plot.ICA}
\title{Plot method for an object of class "ICA".}
\usage{
\method{plot}{ICA}(x, iter = NULL, sensitivity = TRUE,
  print_equivalence = FALSE, ...)
}
\arguments{
\item{x}{an object of class "ICA".}

\item{iter}{iteration. if \code{NULL}, is set to the last iteration.}

\item{sensitivity}{logical; if \code{TRUE} the best design in iteration \code{iter} is checked by equivalence theorem.}

\item{print_equivalence}{logical; if \code{TRUE} the result of checking by equivalence theorem will be printed.
Only applicable when \code{sensitivity = TRUE}.}

\item{...}{argument with no further use.}
}
\description{
Plot method for an object of class "ICA".
}
\seealso{
\code{\link{mica}}, \code{\link{multica_4pl}} and \code{\link{ave}}
}

