% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-UserBayesFunctions.R
\name{print.bayes}
\alias{print.bayes}
\title{Printing \code{bayes} Objects}
\usage{
\method{print}{bayes}(x, iter = NULL, all.info = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{bayes}.}

\item{iter}{Iteration number. if \code{NULL}, it will be set to the last iteration.}

\item{all.info}{Print all the information? Defaults to \code{FALSE}.}

\item{...}{Argument with no further use.}
}
\description{
Print method for an object of class \code{bayes}.
}
\seealso{
\code{\link{bayes}}
}
