% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_print_equivalnce.R
\name{print.equivalence}
\alias{print.equivalence}
\title{print equivalence object}
\usage{
\method{print}{equivalence}(x, ...)
}
\arguments{
\item{x}{an object of class 'equivalence'.}

\item{...}{argument with no further use.}
}
\description{
Print method for an object of class "equivalence".
}
\seealso{
\code{\link{equivalence}}, \code{\link{equivalence_multiple}} and \code{\link{equivalence_on_average}}.
}

