% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_print.R
\name{print.ICA}
\alias{print.ICA}
\title{Print method for an object of class "ICA"}
\usage{
\method{print}{ICA}(x, iter = NULL, ...)
}
\arguments{
\item{x}{an object of class "ICA".}

\item{iter}{iteration.}

\item{...}{argument with no further use.}
}
\description{
Print method for an object of class "ICA"
}
\seealso{
\code{\link{mica}}
}

