% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-UserBayesFunctions.R
\name{print.sensbayes}
\alias{print.sensbayes}
\title{Printing \code{sensbayes} Objects}
\usage{
\method{print}{sensbayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sensbayes}.}

\item{...}{Argument with no further use.}
}
\description{
Print method for an object of class \code{sensbayes}.
}
\seealso{
\code{\link{sensbayes}}, \code{\link{sensbayescomp}}
}
