% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-UserMinimaxFunctions.R
\name{print.sensminimax}
\alias{print.sensminimax}
\title{Printing \code{sensminimax} Objects}
\usage{
\method{print}{sensminimax}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sensminimax}.}

\item{...}{Argument with no further use.}
}
\description{
Print method for an object of class \code{sensminimax}.
}
\seealso{
\code{\link{sensminimax}}, \code{\link{senslocally}}, \code{\link{sensrobust}}
}
