% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FIM_sig_emax}
\alias{FIM_sig_emax}
\title{Fisher Information Matrix for the Sigmoid Emax Model}
\usage{
FIM_sig_emax(x, w, param)
}
\arguments{
\item{x}{Vector of design points.}

\item{w}{Vector of design weight. Its length must be equal to the length of \code{x} and \code{sum(w) = 1}.}

\item{param}{Vector of values for the model parameters \code{c(b1, b2, b3, b4)}.
The mean of response variable is .}
}
\value{
Fisher information matrix.
}
\description{
It provides the cpp function for FIM for the model \code{~b1+(b2-b1)*(x^b4)/(x^b4+b3^b4)}
}
