% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-UserCommonFunctions.R
\name{iterate}
\alias{iterate}
\title{Updating an Object of Class 'bayes' or 'minimax'}
\usage{
iterate(object, iter)
}
\arguments{
\item{object}{An object of class 'minimax' or 'Bayesian'.}

\item{iter}{Number of iterations.}
}
\value{
An (updated) object of class 'bayes' or 'minimax'.
}
\description{
Runs the ICA optimization algorithm on an object of class 'bayes' or 'minimax' for more number of iterations  and updates the results.
}
\seealso{
\code{\link{iterate.minimax}} and \code{\link{iterate.bayes}}.
}
