% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_iterate_methods.R
\name{iterate}
\alias{iterate}
\title{update the object by running the ICA algorithm for more number of iterations.}
\usage{
iterate(object, iter)
}
\arguments{
\item{object}{object of class 'ICA'.}

\item{iter}{number of iterations.}
}
\value{
object of class 'ICA'.
}
\description{
Runs the algorithm for more number of iterations and update the results.
}
\seealso{
\code{\link{iterate.ICA}}
}

