% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-UserMinimaxFunctions.R
\name{print.minimax}
\alias{print.minimax}
\title{Printing \code{minimax} Objects}
\usage{
\method{print}{minimax}(x, iter = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{minimax}.}

\item{iter}{Iteration number. if \code{NULL}, will be set equal to last iteration.}

\item{...}{Argument with no further use.}
}
\description{
Print method for an object of class \code{minimax}.
}
\seealso{
\code{\link{minimax}}, \code{\link{locally}}, \code{\link{robust}}
}
